insightgen-backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Google AI
google-generativeai==0.3.2

# LangChain (for structured prompting)
langchain==0.1.4
langchain-google-genai==0.0.6

# Validation and Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# SQL Parsing and Security
sqlparse==0.4.4

# Utilities
python-dateutil==2.8.2
pandas==2.1.4

# CORS
fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9
